+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Changed
+ - "password" method now adds symbols by default. This can be disabled by setting the "symbols" parameter to False.
+
 ## [1.1.0] - 2023-11-28
 ### Added
 - `MongoDB` support for `database` method.